semi rewrite and rendering code + stats set up
it's a bit spaghetti rn but I'll fix that eventually if I keep working on this project. Also if you set n to as close to 0 as it will go with the controls then it makes a crazy weird ass sound and I'm not entirely sure why. I think it's because the sample rate is too low to detect the chaos of the frequency that it generates.

This now also renders the full wave, the wave on your current frame, and all of the stats for every setting you can edit