string notes + finished rendering stuff
Now I render everything, and all the systems are fucking amazing and so much easier to work with. Controls are easy to implement, and new settings are just as simple. Most of the stuff is commented, and I wrote a simple text class for ease of use when making new text instances. I've also put help functions in globals so that I don't have to worry about clogging up the notesV2.py file.